+++ ElevatedWorker/JunctionPointXP.cs
@@ -1,0 +1,282 @@
+﻿using System;
+using System.IO;
+using System.Runtime.InteropServices;
+using System.Text;
+using Microsoft.Win32.SafeHandles;
+
+namespace ElevatedWorker
+{
+    //Taken from http://www.codeproject.com/KB/files/JunctionPointsNet.aspx
+
+    /// <summary>
+    /// Provides access to NTFS junction points in .Net.
+    /// </summary>
+    public class JunctionPointXp
+    {
+        /// <summary>
+        /// The file or directory is not a reparse point.
+        /// </summary>
+        private const int ERROR_NOT_A_REPARSE_POINT = 4390;
+
+        /// <summary>
+        /// The reparse point attribute cannot be set because it conflicts with an existing attribute.
+        /// </summary>
+        private const int ERROR_REPARSE_ATTRIBUTE_CONFLICT = 4391;
+
+        /// <summary>
+        /// The data present in the reparse point buffer is invalid.
+        /// </summary>
+        private const int ERROR_INVALID_REPARSE_DATA = 4392;
+
+        /// <summary>
+        /// The tag present in the reparse point buffer is invalid.
+        /// </summary>
+        private const int ERROR_REPARSE_TAG_INVALID = 4393;
+
+        /// <summary>
+        /// There is a mismatch between the tag specified in the request and the tag present in the reparse point.
+        /// </summary>
+        private const int ERROR_REPARSE_TAG_MISMATCH = 4394;
+
+        /// <summary>
+        /// Command to set the reparse point data block.
+        /// </summary>
+        private const int FSCTL_SET_REPARSE_POINT = 0x000900A4;
+
+        /// <summary>
+        /// Command to get the reparse point data block.
+        /// </summary>
+        private const int FSCTL_GET_REPARSE_POINT = 0x000900A8;
+
+        /// <summary>
+        /// Command to delete the reparse point data base.
+        /// </summary>
+        private const int FSCTL_DELETE_REPARSE_POINT = 0x000900AC;
+
+        /// <summary>
+        /// Reparse point tag used to identify mount points and junction points.
+        /// </summary>
+        private const uint IO_REPARSE_TAG_MOUNT_POINT = 0xA0000003;
+
+        /// <summary>
+        /// This prefix indicates to NTFS that the path is to be treated as a non-interpreted
+        /// path in the virtual file system.
+        /// </summary>
+        private const string NonInterpretedPathPrefix = @"\??\";
+
+        [DllImport("kernel32.dll", CharSet = CharSet.Auto, SetLastError = true)]
+        private static extern bool DeviceIoControl(IntPtr hDevice, uint dwIoControlCode,
+                                                   IntPtr InBuffer, int nInBufferSize,
+                                                   IntPtr OutBuffer, int nOutBufferSize,
+                                                   out int pBytesReturned, IntPtr lpOverlapped);
+
+        [DllImport("kernel32.dll", SetLastError = true)]
+        private static extern IntPtr CreateFile(
+            string lpFileName,
+            EFileAccess dwDesiredAccess,
+            EFileShare dwShareMode,
+            IntPtr lpSecurityAttributes,
+            ECreationDisposition dwCreationDisposition,
+            EFileAttributes dwFlagsAndAttributes,
+            IntPtr hTemplateFile);
+
+        /// <summary>
+        /// Creates a junction point from the specified directory to the specified target directory.
+        /// </summary>
+        /// <remarks>
+        /// Only works on NTFS.
+        /// </remarks>
+        /// <param name="junctionPoint">The junction point path</param>
+        /// <param name="target">The target directory</param>
+        /// <exception cref="IOException">Thrown when the junction point could not be created or when
+        public void Create(string junctionPoint, string target)
+        {
+            target = Path.GetFullPath(target);
+
+            Directory.CreateDirectory(junctionPoint);
+
+            using (SafeFileHandle handle = OpenReparsePoint(junctionPoint, EFileAccess.GenericWrite))
+            {
+                byte[] targetDirBytes = Encoding.Unicode.GetBytes(NonInterpretedPathPrefix + Path.GetFullPath(target));
+
+                var reparseDataBuffer = new REPARSE_DATA_BUFFER();
+
+                reparseDataBuffer.ReparseTag = IO_REPARSE_TAG_MOUNT_POINT;
+                reparseDataBuffer.ReparseDataLength = (ushort)(targetDirBytes.Length + 12);
+                reparseDataBuffer.SubstituteNameOffset = 0;
+                reparseDataBuffer.SubstituteNameLength = (ushort)targetDirBytes.Length;
+                reparseDataBuffer.PrintNameOffset = (ushort)(targetDirBytes.Length + 2);
+                reparseDataBuffer.PrintNameLength = 0;
+                reparseDataBuffer.PathBuffer = new byte[0x3ff0];
+                Array.Copy(targetDirBytes, reparseDataBuffer.PathBuffer, targetDirBytes.Length);
+
+                int inBufferSize = Marshal.SizeOf(reparseDataBuffer);
+                IntPtr inBuffer = Marshal.AllocHGlobal(inBufferSize);
+
+                try
+                {
+                    Marshal.StructureToPtr(reparseDataBuffer, inBuffer, false);
+
+                    int bytesReturned;
+                    bool result = DeviceIoControl(handle.DangerousGetHandle(), FSCTL_SET_REPARSE_POINT,
+                                                  inBuffer, targetDirBytes.Length + 20, IntPtr.Zero, 0,
+                                                  out bytesReturned, IntPtr.Zero);
+
+                    if (!result)
+                        ThrowLastWin32Error("Unable to create junction point.");
+                }
+                finally
+                {
+                    Marshal.FreeHGlobal(inBuffer);
+                }
+            }
+        }
+
+        private SafeFileHandle OpenReparsePoint(string reparsePoint, EFileAccess accessMode)
+        {
+            var reparsePointHandle = new SafeFileHandle(CreateFile(reparsePoint, accessMode,
+                                                                   EFileShare.Read | EFileShare.Write |
+                                                                   EFileShare.Delete,
+                                                                   IntPtr.Zero, ECreationDisposition.OpenExisting,
+                                                                   EFileAttributes.BackupSemantics |
+                                                                   EFileAttributes.OpenReparsePoint, IntPtr.Zero), true);
+
+            if (Marshal.GetLastWin32Error() != 0)
+                ThrowLastWin32Error("Unable to open reparse point.");
+
+            return reparsePointHandle;
+        }
+
+        private static void ThrowLastWin32Error(string message)
+        {
+            throw new IOException(message, Marshal.GetExceptionForHR(Marshal.GetHRForLastWin32Error()));
+        }
+
+        #region Nested type: ECreationDisposition
+
+        private enum ECreationDisposition : uint
+        {
+            New = 1,
+            CreateAlways = 2,
+            OpenExisting = 3,
+            OpenAlways = 4,
+            TruncateExisting = 5,
+        }
+
+        #endregion
+
+        #region Nested type: EFileAccess
+
+        [Flags]
+        private enum EFileAccess : uint
+        {
+            GenericRead = 0x80000000,
+            GenericWrite = 0x40000000,
+            GenericExecute = 0x20000000,
+            GenericAll = 0x10000000,
+        }
+
+        #endregion
+
+        #region Nested type: EFileAttributes
+
+        [Flags]
+        private enum EFileAttributes : uint
+        {
+            Readonly = 0x00000001,
+            Hidden = 0x00000002,
+            System = 0x00000004,
+            Directory = 0x00000010,
+            Archive = 0x00000020,
+            Device = 0x00000040,
+            Normal = 0x00000080,
+            Temporary = 0x00000100,
+            SparseFile = 0x00000200,
+            ReparsePoint = 0x00000400,
+            Compressed = 0x00000800,
+            Offline = 0x00001000,
+            NotContentIndexed = 0x00002000,
+            Encrypted = 0x00004000,
+            Write_Through = 0x80000000,
+            Overlapped = 0x40000000,
+            NoBuffering = 0x20000000,
+            RandomAccess = 0x10000000,
+            SequentialScan = 0x08000000,
+            DeleteOnClose = 0x04000000,
+            BackupSemantics = 0x02000000,
+            PosixSemantics = 0x01000000,
+            OpenReparsePoint = 0x00200000,
+            OpenNoRecall = 0x00100000,
+            FirstPipeInstance = 0x00080000
+        }
+
+        #endregion
+
+        #region Nested type: EFileShare
+
+        [Flags]
+        private enum EFileShare : uint
+        {
+            None = 0x00000000,
+            Read = 0x00000001,
+            Write = 0x00000002,
+            Delete = 0x00000004,
+        }
+
+        #endregion
+
+        #region Nested type: REPARSE_DATA_BUFFER
+
+        [StructLayout(LayoutKind.Sequential)]
+        private struct REPARSE_DATA_BUFFER
+        {
+            /// <summary>
+            /// Reparse point tag. Must be a Microsoft reparse point tag.
+            /// </summary>
+            public uint ReparseTag;
+
+            /// <summary>
+            /// Size, in bytes, of the data after the Reserved member. This can be calculated by:
+            /// (4 * sizeof(ushort)) + SubstituteNameLength + PrintNameLength + 
+            /// (namesAreNullTerminated ? 2 * sizeof(char) : 0);
+            /// </summary>
+            public ushort ReparseDataLength;
+
+            /// <summary>
+            /// Reserved; do not use. 
+            /// </summary>
+            public readonly ushort Reserved;
+
+            /// <summary>
+            /// Offset, in bytes, of the substitute name string in the PathBuffer array.
+            /// </summary>
+            public ushort SubstituteNameOffset;
+
+            /// <summary>
+            /// Length, in bytes, of the substitute name string. If this string is null-terminated,
+            /// SubstituteNameLength does not include space for the null character.
+            /// </summary>
+            public ushort SubstituteNameLength;
+
+            /// <summary>
+            /// Offset, in bytes, of the print name string in the PathBuffer array.
+            /// </summary>
+            public ushort PrintNameOffset;
+
+            /// <summary>
+            /// Length, in bytes, of the print name string. If this string is null-terminated,
+            /// PrintNameLength does not include space for the null character. 
+            /// </summary>
+            public ushort PrintNameLength;
+
+            /// <summary>
+            /// A buffer containing the unicode-encoded path string. The path string contains
+            /// the substitute name string and print name string.
+            /// </summary>
+            [MarshalAs(UnmanagedType.ByValArray, SizeConst = 0x3FF0)]
+            public byte[] PathBuffer;
+        }
+
+        #endregion
+    }
+}
